DEV-679 Code refactoring based on updated DB structure for service-data
