--- service-data/src/main/java/sg/ncl/service/data/data/jpa/DatasetRepository.java
@@ -1,9 +1,0 @@
-package sg.ncl.service.data.data.jpa;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-/**
- * Created by dcszwang on 10/6/2016.
- */
-public interface DatasetRepository extends JpaRepository<DatasetEntity, String> {
-}
