+++ service-data/src/main/java/sg/ncl/service/data/domain/DataResource.java
@@ -3,8 +3,9 @@ package sg.ncl.service.data.domain;
 /**
  * Created by dcszwang on 10/5/2016.
  */
-@FunctionalInterface
 public interface DataResource {
+
+    Long getId();
 
     String getUri();
