+++ service-data/src/main/java/sg/ncl/service/data/domain/DataResourceType.java
@@ -3,6 +3,6 @@ package sg.ncl.service.data.domain;
 /**
  * Created by dcszwang on 10/5/2016.
  */
-public enum DatasetAccessibility {
-    OPEN, RESTRICTED
+public enum DataResourceType {
+    EXTERNAL, INTERNAL
 }
