+++ service-data/src/main/java/sg/ncl/service/data/domain/DataStatistics.java
@@ -1,0 +1,15 @@
+package sg.ncl.service.data.domain;
+
+import java.time.ZonedDateTime;
+
+/**
+ * Created by dcszwang on 10/7/2016.
+ */
+public interface DataStatistics {
+
+    Long getId();
+
+    String getUserId();
+
+    ZonedDateTime getDate();
+}
