--- service-data/src/main/java/sg/ncl/service/data/domain/DatasetResource.java
@@ -1,13 +1,0 @@
-package sg.ncl.service.data.domain;
-
-/**
- * Created by dcszwang on 10/5/2016.
- */
-public interface DatasetResource {
-
-    String getId();
-
-    String getLink();
-
-    DatasetResourceType getType();
-}
