--- service-data/src/main/java/sg/ncl/service/data/domain/DatasetResourceType.java
@@ -1,8 +1,0 @@
-package sg.ncl.service.data.domain;
-
-/**
- * Created by dcszwang on 10/5/2016.
- */
-public enum DatasetResourceType {
-    EXTERNAL, INTERNAL
-}
