--- service-data/src/main/java/sg/ncl/service/data/domain/DatasetStatus.java
@@ -1,8 +1,0 @@
-package sg.ncl.service.data.domain;
-
-/**
- * Created by dcszwang on 10/5/2016.
- */
-public enum DatasetStatus {
-    IN_PROGRESS, COMPLETE, INACTIVE
-}
