+++ service-data/src/test/java/sg/ncl/service/data/AbstractTest.java
@@ -5,5 +5,5 @@ import org.springframework.boot.test.context.SpringBootTest;
 /**
  * Created by dcszwang on 10/6/2016.
  */
-@SpringBootTest(classes = DatasetApplication.class)
+@SpringBootTest(classes = DataApplication.class)
 public abstract class AbstractTest extends sg.ncl.common.test.AbstractTest {}
